requirements.txt
absl-py==0.11.0
aiogrpc==1.8
alabaster==0.7.12
astunparse==1.6.3
awscli==1.18.216
Babel==2.9.0
bcrypt==3.2.0
blitz-bayesian-pytorch==0.2.7
botocore==1.19.56
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.3
coverage==5.3.1
cycler==0.10.0
dataclasses==0.6
docutils==0.15.2
et-xmlfile==1.0.1
flake8==3.8.4
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
goprocam==4.1.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imagesize==1.2.0
importlib-metadata==3.4.0
imutils==0.5.3
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mavsdk==0.15.0
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==3.0.6
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.0
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.4
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-util==1.2.1
python-utils==2.5.1
pytz==2020.5
PyYAML==5.3.1
qtgui==0.0.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.4
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==2.0.1
zipp==3.4.0
